fix index_compute

+++ csrc/ir_internal_nodes.h
@@ -1244,7 +1244,7 @@ class TORCH_CUDA_CU_API GatherOp : public Expr {
   }
 
   //! Returns the gather axis that corresponds to an input axis
-  int gatherAxis(int axis) const;
+  size_t gatherAxis(size_t axis) const;
 
   //! The size of zero-padding of each axis.
   const auto& padWidth() const {
